Upgrade cocos/math & cocos/network to v3.13

+++ native/cocos/math/Vec3.cpp
@@ -74,7 +74,7 @@ float Vec3::angle(const Vec3& v1, const Vec3& v2)
     float dy = v1.z * v2.x - v1.x * v2.z;
     float dz = v1.x * v2.y - v1.y * v2.x;
 
-    return atan2f(sqrt(dx * dx + dy * dy + dz * dz) + MATH_FLOAT_SMALL, dot(v1, v2));
+    return std::atan2(std::sqrt(dx * dx + dy * dy + dz * dz) + MATH_FLOAT_SMALL, dot(v1, v2));
 }
 
 void Vec3::add(const Vec3& v1, const Vec3& v2, Vec3* dst)
@@ -157,7 +157,7 @@ float Vec3::distance(const Vec3& v) const
     float dy = v.y - y;
     float dz = v.z - z;
 
-    return sqrt(dx * dx + dy * dy + dz * dz);
+    return std::sqrt(dx * dx + dy * dy + dz * dz);
 }
 
 float Vec3::distanceSquared(const Vec3& v) const
@@ -186,7 +186,7 @@ void Vec3::normalize()
     if (n == 1.0f)
         return;
 
-    n = sqrt(n);
+    n = std::sqrt(n);
     // Too close to zero.
     if (n < MATH_TOLERANCE)
         return;
@@ -228,4 +228,3 @@ const Vec3 Vec3::UNIT_Y(0.0f, 1.0f, 0.0f);
 const Vec3 Vec3::UNIT_Z(0.0f, 0.0f, 1.0f);
 
 NS_CC_MATH_END
-
